requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
joblib
tensorflow
rapidfuzz
tldextract
python-whois
fpdf2